i18n_macros

Macro attr

attr!() { /* proc-macro */ }

Queries an attribute from an existing i18n::Message instance.

This macro provides a convenient way to lazily query for an attribute’s value from a Message struct that you already have. If the query fails, it will invoke the on_error handler on the Locales instance.

Syntax

attr!(message_expr, attr_id: LitStr [, locales = VAR_NAME] [, key = value]*)

message_expr: An expression that evaluates to a &mut i18n::Message.

attr_id: A string literal representing the ID of the attribute to query.

locales (optional): The i18n::Locales instance to use for the error callback. Defaults to LOCALES.

key = value: Optional key-value pairs for arguments to the attribute.

Returns

A String. If the attribute doesn’t exist or an error occurs during formatting, it returns the attribute ID as a fallback string.

Example

i18n::load!("../tests/i18n", name = ATTR_MACRO_LOCALES);
let lang = i18n::langid!("en-US");

let mut msg = i18n::tr!(lang, "login-btn", locales = ATTR_MACRO_LOCALES);

// Get an attribute with no arguments (using default LOCALES)
let idle_text = i18n::attr!(msg, "idle");
assert_eq!(idle_text, "Login");

// Get an attribute with arguments and a custom locales instance
let arg_text = i18n::attr!(msg, "attr-arg", locales = ATTR_MACRO_LOCALES, "text" = "some value");
assert_eq!(arg_text, "This is an attribute argument with arbitrary text: \u{2068}some value\u{2069}");