i18n_macros

Macro load

load!() { /* proc-macro */ }

Loads Fluent localization files from a specified directory and creates a lazily-initialized static instance of i18n::Locales.

This macro reads .ftl files from subdirectories of the given path (each subdirectory representing a locale). It parses these files at compile time and embeds their content into your binary.

The generated static instance provides methods to query for localized messages.

Syntax

load!(path: LitStr [, fallback_lang: LitStr] [, check_keys: bool] [, name: Ident] [, on_error: Expr])

Arguments

path: A string literal representing the path to the locales directory. This path should be relative to your crate root (where Cargo.toml is).

fallback_lang (optional): A string literal representing the language identifier (e.g., “en-US”) to use as a fallback if a message is not found in the requested language. Defaults to "en-US".

check_keys (optional): A boolean literal (true or false). If true (default), the macro will perform a compile-time check to ensure all locale files have a consistent set of message keys. If false, this check is skipped.

name (optional): An identifier to use as the name for the generated lazy_static variable. Defaults to LOCALES.

on_error (optional): An expression that evaluates to a function or closure to be called when an error occurs during localization (e.g., missing message). The function should have the signature fn(errors: &[i18n::FluentError]).

Usage

// Basic usage with default values.
i18n::load!("../tests/i18n", name = LOCALES_DEFAULT);

// With a custom fallback language, disabled key checking, and a custom name.
i18n::load!(
    "../tests/i18n",
    fallback_lang = "hr-hr",
    check_keys = false,
    name = MY_I18N_DATA
);

// With an error handler.
fn on_error(errors: &[i18n::FluentError]) {
    // Log the error, send it to a monitoring service, etc.
    println!("Localization errors: {:?}", errors);
}

i18n::load!("../tests/i18n", on_error = on_error, name = LOCALES_WITH_ERROR_HANDLER);

// Example of accessing the generated data (assuming default name `LOCALES_DEFAULT`).
let lang_en = i18n::langid!("en-US");
let query = i18n::Query::new("foo-a");
let message = LOCALES_DEFAULT.query(&lang_en, &query).unwrap();
assert_eq!(message.value, "English A".to_string());

let lang_hr = i18n::langid!("hr-hr");
let query = i18n::Query::new("foo-a");
let message = LOCALES_DEFAULT.query(&lang_hr, &query).unwrap();
assert_eq!(message.value, "Croatian A".to_string());